+++ integrations/github/src/index.ts
@@ -1,4 +1,4 @@
-import { Router } from 'itty-router';
+import { Router, error } from 'itty-router';
 
 import { ContentKitIcon, ContentKitSelectOption, GitSyncOperationState } from '@gitbook/api';
 import {
@@ -21,8 +21,9 @@ import {
 import { configBlock } from './components';
 import { getGitHubAppJWT } from './provider';
 import { triggerExport, updateCommitWithPreviewLinks } from './sync';
-import type { GithubRuntimeContext } from './types';
-import { BRANCH_REF_PREFIX } from './utils';
+import { handleIntegrationTask } from './tasks';
+import type { GithubRuntimeContext, IntegrationTask } from './types';
+import { arrayToHex, BRANCH_REF_PREFIX, safeCompare } from './utils';
 import { handlePullRequestEvents, handlePushEvent, verifyGitHubWebhookSignature } from './webhooks';
 
 const logger = Logger('github');
@@ -36,6 +37,61 @@ const handleFetchEvent: FetchEventCallback<GithubRuntimeContext> = async (reques
                 environment.installation?.urls.publicEndpoint ||
                 environment.integration.urls.publicEndpoint
         ).pathname,
+    });
+
+    async function verifyIntegrationSignature(
+        payload: string,
+        signature: string,
+        secret: string
+    ): Promise<boolean> {
+        if (!signature) {
+            return false;
+        }
+
+        const algorithm = { name: 'HMAC', hash: 'SHA-256' };
+        const enc = new TextEncoder();
+        const key = await crypto.subtle.importKey('raw', enc.encode(secret), algorithm, false, [
+            'sign',
+            'verify',
+        ]);
+        const signed = await crypto.subtle.sign(algorithm.name, key, enc.encode(payload));
+        const expectedSignature = arrayToHex(signed);
+
+        return safeCompare(expectedSignature, signature);
+    }
+
+    /**
+     * Handle integration tasks
+     */
+    router.post('/tasks', async (request) => {
+        const signature = request.headers.get('x-gitbook-integration-signature') ?? '';
+        const payloadString = await request.text();
+
+        const verified = await verifyIntegrationSignature(
+            payloadString,
+            signature,
+            environment.signingSecret!
+        );
+
+        if (!verified) {
+            return new Response('Invalid integration signature', {
+                status: 400,
+            });
+        }
+
+        const { task } = JSON.parse(payloadString) as { task: IntegrationTask };
+        logger.debug('verified & received integration task', task);
+
+        context.waitUntil(
+            (async () => {
+                await handleIntegrationTask(context, task);
+            })()
+        );
+
+        return new Response(JSON.stringify({ acknowledged: true }), {
+            status: 200,
+            headers: { 'content-type': 'application/json' },
+        });
     });
 
     /**
@@ -309,15 +365,10 @@ const handleFetchEvent: FetchEventCallback<GithubRuntimeContext> = async (reques
         });
     });
 
-    let response;
-    try {
-        response = await router.handle(request, context);
-    } catch (error: any) {
-        logger.error('error handling request', error);
-        return new Response(error.message, {
-            status: error.status || 500,
-        });
-    }
+    const response = (await router.handle(request, context).catch((err) => {
+        logger.error('error handling request', err);
+        return error(err);
+    })) as Response | undefined;
 
     if (!response) {
         return new Response(`No route matching ${request.method} ${request.url}`, {
@@ -349,11 +400,18 @@ const handleSpaceContentUpdated: EventCallback<
 
     const spaceInstallation = context.environment.spaceInstallation;
     if (!spaceInstallation) {
-        logger.debug(`missing space installation, skipping`);
+        logger.debug(`missing space installation for ${event.spaceId}, skipping`);
         return;
     }
 
-    await triggerExport(context, spaceInstallation);
+    if (!spaceInstallation.configuration.key) {
+        logger.debug(`space ${event.spaceId} is not configured, skipping`);
+        return;
+    }
+
+    await triggerExport(context, spaceInstallation, {
+        eventTimestamp: new Date(revision.createdAt),
+    });
 };
 
 /*